backend/requirements.txt
# Core Django
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-environ==0.11.2

# Database
psycopg2-binary==2.9.10
dj-database-url==2.2.0

# Authentication & Security
PyJWT==2.9.0
jwcrypto==1.5.6
bcrypt==4.2.1
firebase-admin==6.6.0

# Celery & Redis
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.0

# API Documentation
drf-yasg==1.21.8

# Production Server
gunicorn==23.0.0
whitenoise==6.8.2

# AI & External APIs
google-generativeai==0.8.3
google-cloud-storage==2.19.0
language-tool-python==2.8.1
wikipedia==1.4.0
yfinance==0.2.50
PyGithub==2.5.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Document Generation
python-docx==1.1.2
reportlab==4.2.5
Jinja2==3.1.4

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
coverage==7.6.9
faker==33.1.0
factory-boy==3.3.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
